+++ src/main/java/com/fiap/tech/application/controllers/admin/pedidos/AdminGetPedidosController.java
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("admin/pedido")
-public class GetPedidosController {
+public class AdminGetPedidosController {
 
     private final PedidoRepository pedidoRepository;
     private final PedidoProdutoRepository pedidoProdutoRepository;
